move interfaces from index to its own file
